pyproject.toml
[tool.poetry]
name = "traffic-example"
version = "0.0.1"
description = ""
authors = ["alex.tarasiuk"]
readme = "README.md"
packages = [{ include = "tra" }]

[tool.poetry.dependencies]
python = "^3.12"
dependency-injector = "4.42.0b1"
fastapi = "^0.112.1"
pydantic = "^2.8.2"
python-ulid = {extras = ["pydantic"], version = "^2.7.0"}
pymongo = "^4.8.0"


[tool.poetry.group.local.dependencies]
pre-commit = "^3.8.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
faker = "^27.0.0"
factory-boy = "^3.3.0"
coverage = "^7.6.1"
mypy = "^1.11.1"


[tool.poetry.group.production.dependencies]
gunicorn = "^23.0.0"
uvicorn = "^0.30.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
src_paths = ["tra", "tests"]
skip = ["migrations"]
multi_line_output = 3
include_trailing_comma = true
line_length = 120

[tool.black]
line-length = 120

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
python_files = "tests.py test_*.py *_tests.py"

[tool.mypy]
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
#
strict_equality = true
extra_checks = true
#
check_untyped_defs = true
#
disallow_any_generics = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
disallow_untyped_defs = true

[tool.autoflake]
remove_all_unused_imports = true
ignore_init_module_imports = true
files = [
    "tra/*.py",
    "tests/*.py"
]

[tool.mypy-pymongo]
follow_imports = false
